fix convert_image_blocks to work with ExhibitPage
there were references to `page.body` that would only work for pages
with a body field in their model, but ExhibitPage's streamfield is
named description. I used setattr() in one place to work around this
but not everywhere it was needed. I actually think there are no image
blocks in ExhibitPages anyways but it's still good to fix this.